# FastMod for EEHashTable (Faster virtual generics)