Closes https://github.com/dotnet/runtime/issues/65778

It was suggested to use a different data structure for `CORINFO_HELP_VIRTUAL_FUNC_PTR` but that is a more complicated change and I think this one worth having too anyway.

Benchmark:
```csharp
using System.Text.Json.Nodes;
using BenchmarkDotNet.Attributes;
using BenchmarkDotNet.Running;

BenchmarkSwitcher.FromAssembly(typeof(Benchmarks).Assembly).Run(args);

public class Benchmarks
{
    public virtual void VGM<T>() { }

    [Benchmark]
    public void CallVirtualGenericMethod() => VGM<int>();



    // System.Text.Json
    JsonNode _jsonNode = true;

    [Benchmark]
    public bool JsonNodeConversion() => (bool)_jsonNode;



    // string interning
    string str = "Hello";

    [Benchmark]
    public string StrIsInterned() => string.IsInterned(str);
}
```

|                    Method |                 Toolchain |      Mean |     Error |    StdDev | Ratio |
|-------------------------- |-------------------------- |----------:|----------:|----------:|------:|
|  CallVirtualGenericMethod |    \Core_Root\corerun.exe |  5.299 ns | 0.0364 ns | 0.0322 ns |  1.00 |
|  CallVirtualGenericMethod | \Core_Root_PR\corerun.exe |  4.166 ns | 0.1064 ns | 0.1383 ns |  0.79 |
|                           |                           |           |           |           |       |
|        JsonNodeConversion |    \Core_Root\corerun.exe |  5.509 ns | 0.0070 ns | 0.0062 ns |  1.00 |
|        JsonNodeConversion | \Core_Root_PR\corerun.exe |  4.177 ns | 0.0461 ns | 0.0431 ns |  0.76 |
|                           |                           |           |           |           |       |
|             StrIsInterned |    \Core_Root\corerun.exe | 54.741 ns | 0.1900 ns | 0.1780 ns |  1.00 |
|             StrIsInterned | \Core_Root_PR\corerun.exe | 50.172 ns | 0.3521 ns | 0.2750 ns |  0.92 |