@VSadov 2022-02-27
In the long term we should experiment with a more lightweight caching scheme for virtual generics. 
However, this change is an obvious improvement to the current implementation and will benefit everything that uses `EEHashTable`.