@elinor-fung 2022-03-01
@EgorBo `System.Runtime.InteropServices.Tests.GetEndComSlotTests.GetEndComSlot_Windows_ReturnsExpected` are failing on x64 with this change: https://dev.azure.com/dnceng/public/_build/results?buildId=1634854&view=ms.vss-test-web.build-test-results-tab&runId=45284702&resultId=151347&paneView=debug

```
System.Runtime.InteropServices.Tests.GetEndComSlotTests.GetEndComSlot_Windows_ReturnsExpected(type: typeof(System.Runtime.InteropServices.Tests.Common.IComImportObject), expected: 6) [FAIL]
  System.DivideByZeroException : Attempted to divide by zero.
  Stack Trace:
       at System.Runtime.InteropServices.Marshal.GetEndComSlot(Type t)
```